refactor: products와 shoppingCart 관련 타입 및 폴더 구조 정리
