+++ src/APIs/products/type.ts
@@ -1,0 +1,12 @@
+import { Product } from "../../types/product.type";
+
+interface FetchProductsRequest {
+  endpoint: string;
+}
+
+interface FetchProductsResponse {
+  content: Product[];
+  totalPages: number;
+}
+
+export type { FetchProductsRequest, FetchProductsResponse };
