--- src/components/ProductListToolBar/toolBar.constant.ts
@@ -1,4 +1,0 @@
-const CATEGORY = Object.freeze(["전체", "패션잡화", "식료품"]);
-const SORT_OPTION = Object.freeze(["낮은 가격순", "높은 가격순"]);
-
-export { CATEGORY, SORT_OPTION };
