+++ src/types/product.type.ts
@@ -1,7 +1,5 @@
-import {
-  SORT_OPTION,
-  CATEGORY,
-} from "../components/ProductListToolBar/toolBar.constant";
+export const CATEGORY = Object.freeze(["전체", "패션잡화", "식료품"]);
+export const SORT_OPTION = Object.freeze(["낮은 가격순", "높은 가격순"]);
 
 type Category = (typeof CATEGORY)[number];
 type SortOption = (typeof SORT_OPTION)[number];
@@ -14,62 +12,10 @@ interface Product {
   category: Category;
 }
 
-interface FetchProductsResponse {
-  content: Product[];
-  totalPages: number;
-}
-
 interface CartItem {
   id: number;
   quantity: number;
   product: Product;
-}
-
-interface ProductListResponse {
-  content: Product[];
-  pageable: Pageable;
-  totalElements: number;
-  totalPages: number;
-  last: boolean;
-  size: number;
-  number: number;
-  sort: Sort;
-  numberOfElements: number;
-  first: boolean;
-  empty: boolean;
-}
-interface Pageable {
-  pageNumber: number;
-  pageSize: number;
-  sort: Sort;
-  offset: number;
-  paged: boolean;
-  unpaged: boolean;
-}
-
-interface Sort {
-  empty: boolean;
-  sorted: boolean;
-  unsorted: boolean;
-}
-
-interface FetchProductsRequest {
-  endpoint: string;
-}
-
-interface ShoppingCartResponse {
-  endpoint: string;
-  requestBody?: PostRequestBody | PatchRequestBody;
-  cartItemId?: number;
-}
-
-interface PostRequestBody {
-  productId: number;
-  quantity: number;
-}
-
-interface PatchRequestBody {
-  quantity: number;
 }
 
 interface Error {
@@ -77,16 +23,4 @@ interface Error {
   errorMessage: string;
 }
 
-export type {
-  Product,
-  FetchProductsRequest,
-  ProductListResponse,
-  Pageable,
-  Sort,
-  Category,
-  SortOption,
-  ShoppingCartResponse,
-  CartItem,
-  Error,
-  FetchProductsResponse,
-};
+export type { Product, Category, SortOption, CartItem, Error };
